make base class separate file
