+++ src/CMakeLists.txt
@@ -232,6 +232,7 @@ add_library(parthenon
   solvers/cg_solver_stages.hpp
   solvers/mg_solver.hpp
   solvers/mg_solver_stages.hpp
+  solvers/solver_base.hpp
   solvers/solver_utils.hpp
   solvers/solver_utils_stages.hpp
