+++ src/solvers/bicgstab_solver_stages.hpp
@@ -25,6 +25,7 @@
 #include "kokkos_abstraction.hpp"
 #include "solvers/mg_solver.hpp"
 #include "solvers/mg_solver_stages.hpp"
+#include "solvers/solver_base.hpp"
 #include "solvers/solver_utils_stages.hpp"
 #include "tasks/tasks.hpp"
 #include "utils/type_list.hpp"
